fix DataBase Connection + Mappers

+++ Transactions/ReadTransactionsWallets.Domain/Model/Database/RunTimeController.cs
@@ -5,7 +5,7 @@ namespace ReadTransactionsWallets.Domain.Model.Database
     public class RunTimeController : Entity
     {
         public int? IdRuntime { get; set; }
-        public int? UnixTimeSeconds { get; set; }
+        public decimal? UnixTimeSeconds { get; set; }
         public bool? IsRunning { get; set; }
     }
 }
